--- a/Assets/UnZENity-Core/Resources/Prefabs/Menus/LoadingArea.prefab.meta
+++ b/Assets/UnZENity-Core/Resources/Prefabs/Menus/LoadingArea.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 03e7b79559c84ac4385aae4c205ba41d
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: