--- a/Assets/UnZENity-Core/Resources/Prefabs/Vobs/Default.prefab.meta
+++ b/Assets/UnZENity-Core/Resources/Prefabs/Vobs/Default.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e1c9529ac2f816447bee24487fbe4a4f
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: