--- a/Assets/UnZENity-Core/Scenes/Bootstrap.unity
+++ b/Assets/UnZENity-Core/Scenes/Bootstrap.unity
@@ -218,6 +218,7 @@ GameObject:
   - component: {fileID: 583338022}
   - component: {fileID: 583338021}
   - component: {fileID: 583338020}
+  - component: {fileID: 583338023}
   m_Layer: 0
   m_Name: GameManager
   m_TagString: Untagged
@@ -311,6 +312,18 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   _music: {fileID: 1283201087}
+--- !u!114 &583338023
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 583338016}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: a468d73927dc467392663c8a089bc648, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!1 &1283201085
 GameObject:
   m_ObjectHideFlags: 0