--- a/Assets/UnZENity-Core/Scenes/PreCaching.unity
+++ b/Assets/UnZENity-Core/Scenes/PreCaching.unity
@@ -122,7 +122,7 @@ NavMeshSettings:
     debug:
       m_Flags: 0
   m_NavMeshData: {fileID: 0}
---- !u!1 &315427610
+--- !u!1 &1781062260
 GameObject:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -130,8 +130,8 @@ GameObject:
   m_PrefabAsset: {fileID: 0}
   serializedVersion: 6
   m_Component:
-  - component: {fileID: 315427612}
-  - component: {fileID: 315427611}
+  - component: {fileID: 1781062262}
+  - component: {fileID: 1781062263}
   m_Layer: 0
   m_Name: Bootstrap
   m_TagString: Untagged
@@ -139,26 +139,13 @@ GameObject:
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
   m_IsActive: 1
---- !u!114 &315427611
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 315427610}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 8cb88057a76d4858a021b773245b00a6, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  _loadingArea: {fileID: 912042971}
---- !u!4 &315427612
+--- !u!4 &1781062262
 Transform:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 315427610}
+  m_GameObject: {fileID: 1781062260}
   serializedVersion: 2
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
   m_LocalPosition: {x: 0, y: 0, z: 0}
@@ -167,12 +154,20 @@ Transform:
   m_Children: []
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &912042971 stripped
-GameObject:
-  m_CorrespondingSourceObject: {fileID: 2275007553886627835, guid: 03e7b79559c84ac4385aae4c205ba41d, type: 3}
-  m_PrefabInstance: {fileID: 7634307503045520803}
+--- !u!114 &1781062263
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
---- !u!1001 &7634307503045520803
+  m_GameObject: {fileID: 1781062260}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 850d43d6d77c4099808ba3daf62885eb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  _loadingArea: {fileID: 2085289318}
+--- !u!1001 &2085289317
 PrefabInstance:
   m_ObjectHideFlags: 0
   serializedVersion: 2
@@ -229,9 +224,14 @@ PrefabInstance:
     m_AddedGameObjects: []
     m_AddedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 03e7b79559c84ac4385aae4c205ba41d, type: 3}
+--- !u!1 &2085289318 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 2275007553886627835, guid: 03e7b79559c84ac4385aae4c205ba41d, type: 3}
+  m_PrefabInstance: {fileID: 2085289317}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1660057539 &9223372036854775807
 SceneRoots:
   m_ObjectHideFlags: 0
   m_Roots:
-  - {fileID: 315427612}
-  - {fileID: 7634307503045520803}
+  - {fileID: 1781062262}
+  - {fileID: 2085289317}