--- a/Assets/UnZENity-Core/Scenes/PreCaching.unity.meta
+++ b/Assets/UnZENity-Core/Scenes/PreCaching.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 7e75efa417ad914418311d5d24258d42
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: