--- a/Assets/UnZENity-Core/Scripts/Creator/Meshes/V2/Builder/TextureArrayBuilder.cs.meta
+++ b/Assets/UnZENity-Core/Scripts/Creator/Meshes/V2/Builder/TextureArrayBuilder.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: f22cbe53f00d452bba0e32fc4e8b16e7
-timeCreated: 1710831863