--- a/Assets/UnZENity-Core/Scripts/Extensions/StopwatchExtension.cs.meta
+++ b/Assets/UnZENity-Core/Scripts/Extensions/StopwatchExtension.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 3d649b15e4f147f282b1c0cb23f527ea
+timeCreated: 1727785891