--- a/Assets/UnZENity-Core/Scripts/Manager/CacheManager.cs.meta
+++ b/Assets/UnZENity-Core/Scripts/Manager/CacheManager.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7d32a9dfa18943a595b62f34fedd15b2
+timeCreated: 1727687644