@JaXt0r 2024-10-04
Some updates on the PreCaching logic:
* We save 2 seconds of loading time compared to previous logic of loading. 😕
* I can still optimize Phase 2 a little, but I don't expect huge frame benefits

This is the data for now:

**old logic without cache** 
* 26s/27s - full world + VObs loading - (but unoptimized for fps - aka there are hickups and if optimized it would take about 30-35 I assume)

**new logic with cache**
* 24/25 - full world + VOB loading
* ->  7s - Phase 1 - Cache loading from file and apply Meshes to scene
* -> 12s - Phase 2 - Apply prefab to cached VOB meshes - Prefabs applied to 11.286 VOBs
* ->  6s - Phase 3 - Create dynamic/non-meshed VOBs - created new VOBs 3.267

**Some points against current solution:**
* The most loading time consuming tasks are the application of GameObject components and hierarchies from Pre-cached VOBs and the Prefabs. This is heavily used already and consumes most of the time as we apply it to 11k objects when loaded.
* Precaching VOBs is a huge complexity addition where we now have 3 required phases of Component additions to VOBs
* The more we put into the cache, the more likely we will miss to keep track of when a cache needs to be altered or not. (Cache version flag)
* Loading fifteen thousand objects when we will see only a handful of them during game sessions feels so wrong

## Next steps:
* The Branch behind this PR can be referenced for a couple of weeks if needed
* A possibility could be to only cache TextureArray information and Meshes. No VOB object structure, but instead applying the meshes at the time of world loading.
* Some smaller improvements (like disabling meshes during loading time) are currently adopted to main #154